Relocate tests from framework/events/events.go & update polling check
